Refactored workload api client
