+++ mqtt/mqtt-edgehub/src/edgelet/workload/mod.rs
@@ -1,0 +1,5 @@
+mod client;
+mod models;
+
+pub use client::{WorkloadClient, WorkloadError};
+pub use models::*;
